实现UI调参、直立、行走、转向、结合视觉数据进行寻迹

+++ Software/Stm32f407vgt6/Application/pid.h
@@ -60,8 +60,8 @@ typedef struct _PID_TypeDef
 	
 	void (*f_param_init)(volatile struct _PID_TypeDef *pid,  //PID参数初始化
 				PID_ID id,
-				uint16_t maxOutput,
-				uint16_t integralLimit,
+				float maxOutput,
+				float integralLimit,
 				float deadband,
 				float controlPeriod,    
 				float  target,
